requirements.txt
Pillow
requests
Flask
gunicorn
